fix translations and config
