+++ src/resources/i18n/fr-FR/translation.json
@@ -8,8 +8,8 @@
   "Iniciar sesión": "Log In",
   "Usuario registrado con éxito": "User successfully registered",
   "Registrarse": "Sign Up",
-  "Tarea": "MISSING_TRANSLATION",
-  "Cuaderno": "MISSING_TRANSLATION",
+  "Tarea": "Task",
+  "Cuaderno": "Notebook",
   "Digital": "Digital",
   "Bienvenido a tu cuaderno digital": "Welcome to your digital notebook",
   "Lleva tu cuaderno en la palma de tu mano. Agrega, organiza y accede a tus notas de forma fácil y rápida con nuestra página web de Cuaderno Digital. Toma el control de tus ideas estés donde estés": "Take your notebook in the palm of your hand. Easily add, organize, and access your notes quickly with our Digital Notebook website. Take control of your ideas wherever you are."
